--- src/views/uBikeTable/components/pagination.vue
@@ -1,48 +1,0 @@
-<script setup>
-const props = defineProps({
-    pagerEnd: Number,
-    pagerAddAmount: Number,
-    currentPage: Number,
-    totalPageCount: Number,
-})
-
-const emit = defineEmits(['setPage']);
-
-// 換頁
-const setPage = (page) => {
-  if (page < 1 || page > props.totalPageCount) return;
-  emit('setPage', page); 
-};
-</script>
-
-<template>
-    <nav v-if="pagerEnd > 0">
-      <ul class="pagination">
-        <li @click.prevent="setPage(1)" class="page-item">
-          <a class="page-link" href>第一頁</a>
-        </li>
-        <li @click.prevent="setPage(currentPage - 1)" class="page-item">
-          <a class="page-link" href>&lt;</a>
-        </li>
-  
-        <li v-for="i in pagerEnd" :class="{ active: i + pagerAddAmount === currentPage }" :key="i"
-          @click.prevent="setPage(i + pagerAddAmount)" class="page-item">
-          <a class="page-link" href>{{ i + pagerAddAmount }}</a>
-        </li>
-  
-        <li @click.prevent="setPage(currentPage + 1)" class="page-item">
-          <a class="page-link" href>&gt;</a>
-        </li>
-        <li @click.prevent="setPage(totalPageCount)" class="page-item">
-          <a class="page-link" href>最末頁</a>
-        </li>
-      </ul>
-    </nav>
-  </template>
-
-<style scoped>
-.pagination {
-  display: flex;
-  justify-content: center;
-}
-</style>
