--- src/views/uBikeTable/components/search.vue
@@ -1,9 +1,0 @@
-<script setup>
-const searchTextVal = defineModel('searchText');
-
-</script>
-<template>
-    <p class="mb-3">
-    站點名稱搜尋: <input class="border" type="text" v-model="searchTextVal">
-    </p>
-</template>
