+++ src/views/uBikeTable/uBikeTable.vue
@@ -1,9 +1,5 @@
 <script setup>
 import { ref, computed, watch } from 'vue';
-//import  {pagination} from '../components/pagination.vue'
-import search from './components/search.vue'
-import uBikeTable from './components/uBikeTable.vue'
-import pagination from './components/pagination.vue'
 // 修改這份 YouBike 即時資訊表，並加上
 // 1. 站點名稱搜尋
 // 2. 目前可用車輛 / 總停車格 的排序功能
@@ -114,15 +110,91 @@ const setSort = sortType => {
     isSortDesc.value = false;
   }
 };
+
+// 關鍵字 Highlight
+const keywordsHighlight = (text, keyword) => {
+  if(keyword === '') return text;
+  const reg = new RegExp(keyword, 'gi');
+  return text.replace(reg, `<span style="color: red;">${keyword}</span>`);
+};
 </script>
 
 <template>
   <div class="app">
-    <search @update:searchText="(childMsg) => searchText = childMsg"/>
+    <p class="mb-3">
+      站點名稱搜尋: <input class="border" type="text" v-model="searchText">
+    </p>
+
+    <table class="table table-striped">
+      <thead>
+        <tr>
+          <th @click="setSort('sno')">
+            #
+            <span v-show="currentSort === 'sno'">
+              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
+            </span>
+          </th>
+          <th>
+            場站名稱
+          </th>
+          <th>
+            場站區域
+          </th>
+          <th @click="setSort('available_return_bikes')" class="pointer">
+            目前可用車輛
+            <span v-show="currentSort === 'available_return_bikes'">
+              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
+            </span>
+          </th>
+          <th @click="setSort('total')" class="pointer">
+            總停車格
+            <span v-show="currentSort === 'total'">
+              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
+            </span>
+          </th>
+          <th>
+            資料更新時間
+          </th>
+        </tr>
+      </thead>
+      <tbody>
+        <!-- 替換成 slicedUbikeStops -->
+        <tr v-for="s in slicedUbikeStops" :key="s.sno">
+          <td>{{ s.sno }}</td>
+          <td v-html="keywordsHighlight(s.sna, searchText)"></td>
+          <td>{{ s.sarea }}</td>
+          <td>{{ s.available_return_bikes }}</td>
+          <td>{{ s.total }}</td>
+          <td>{{ (s.mday) }}</td>
+        </tr>
+      </tbody>
+    </table>
   </div>
-  <uBikeTable :currentSort="currentSort" :slicedUbikeStops="slicedUbikeStops" :searchText="searchText" @setSort="setSort"/>
+
   <!-- 頁籤 -->
-  <pagination :pagerEnd="pagerEnd" :pagerAddAmount="pagerAddAmount" :currentPage="currentPage" :totalPageCount="totalPageCount" @setPage="setPage"/>
+  <nav v-if="pagerEnd > 0">
+    <ul class="pagination">
+
+      <li @click.prevent="setPage(1)" class="page-item">
+        <a class="page-link" href>第一頁</a>
+      </li>
+      <li @click.prevent="setPage(currentPage - 1)" class="page-item">
+        <a class="page-link" href>&lt;</a>
+      </li>
+
+      <li v-for="i in pagerEnd" :class="{ active: i + pagerAddAmount === currentPage }" :key="i"
+        @click.prevent="setPage(i + pagerAddAmount)" class="page-item">
+        <a class="page-link" href>{{ i + pagerAddAmount }}</a>
+      </li>
+
+      <li @click.prevent="setPage(currentPage + 1)" class="page-item">
+        <a class="page-link" href>&gt;</a>
+      </li>
+      <li @click.prevent="setPage(totalPageCount)" class="page-item">
+        <a class="page-link" href>最末頁</a>
+      </li>
+    </ul>
+  </nav>
 </template>
 
 <style lang="scss" scoped>
@@ -132,6 +204,11 @@ const setSort = sortType => {
 
 .pointer {
   cursor: pointer;
+}
+
+.pagination {
+  display: flex;
+  justify-content: center;
 }
 
 @media (max-width: 768px) {
